refactor(core): split streaming.rs tests into a sibling module (#3791)
The scan-report additions took streaming.rs to 411 lines, over the
module-size ratchet's 400-line limit and with no allowlist row, so
`no_module_grows_past_its_ratchet_budget` failed and the branch could
not merge.

The production half of the file is 290 lines, comfortably under the
rule; the overflow is the inline `#[cfg(test)] mod tests`, which the
ratchet itself exempts when it lives in a `*_tests.rs` sibling. So the
split follows the crate convention already used by georef.rs,
columnar_index.rs, units.rs and fast_parse.rs rather than carving up
production code that is not actually oversized.

All four tests still run, under `streaming::streaming_tests::*`.
